requirements.txt
UploadFile
fastapi==0.75.2
opencv-python-headless==4.5.4.60
Pillow==8.4.0
pydantic==1.9.0
pyinstaller==5.0.1
starlette==0.17.1
uvicorn==0.17.6
gunicorn


